zmq4: preserve dialed ipc socket when closing
When closing a dialed ipc socket, do not remove the socket file
since we are not the server. Removing the unix socket file on disk
only if net.Listen was used to create the file.

+++ socket.go
@@ -110,7 +110,9 @@ func (sck *socket) Close() error {
 			err = e
 		}
 	}
-	if strings.HasPrefix(sck.ep, "ipc://") {
+
+	// Remove the unix socket file if created by net.Listen
+	if sck.listener != nil && strings.HasPrefix(sck.ep, "ipc://") {
 		os.Remove(sck.ep[len("ipc://"):])
 	}
